refactor(orchestration): modularize coordinator into preparation, execution, and types
