+++ tests/integration/test_render_pipeline.py
@@ -3,7 +3,9 @@ from pathlib import Path
 import pytest
 from PIL import Image
 
-from frame_compare.render import apply_overlay, render_batch, render_frame
+from frame_compare.render.encoders import render_frame
+from frame_compare.render.orchestrator import render_batch
+from frame_compare.render.overlay import apply_overlay
 from frame_compare.render.types import (
     EncoderSettings,
     OverlayConfig,
