+++ tests/orchestration/test_ffmpeg_overlay_application.py
@@ -13,7 +13,8 @@ from frame_compare.orchestration.context import (
     ClipState,
     RunContext,
 )
-from frame_compare.orchestration.coordinator import _run_render_phase, _RunArtifacts
+from frame_compare.orchestration.execution import run_render_phase
+from frame_compare.orchestration.types import RunArtifacts
 from frame_compare.utils.types import WorkspacePaths
 
 
@@ -77,10 +78,10 @@ def test_ffmpeg_extraction_applies_overlay_post_process(
         reporter=None,
     )
 
-    artifacts = _RunArtifacts()
+    artifacts = RunArtifacts()
     runner = FakeFFmpegRunner()
 
-    _run_render_phase(
+    run_render_phase(
         ctx=ctx,
         frames=[0, 1],
         runner=runner,
