+++ tests/orchestration/test_overlay_selection_label_domain.py
@@ -14,7 +14,8 @@ from frame_compare.orchestration.context import (
     ClipState,
     RunContext,
 )
-from frame_compare.orchestration.coordinator import _run_render_phase, _RunArtifacts
+from frame_compare.orchestration.execution import run_render_phase
+from frame_compare.orchestration.types import RunArtifacts
 from frame_compare.utils.types import WorkspacePaths
 
 
@@ -76,9 +77,9 @@ def test_selection_labels_are_looked_up_in_reference_source_frame_domain_after_t
         selection_breakdown=SelectionBreakdown(quantile_dark=[10]),
     )
 
-    artifacts = _RunArtifacts()
+    artifacts = RunArtifacts()
 
-    _run_render_phase(
+    run_render_phase(
         ctx=ctx,
         frames=[0],
         runner=FakeFFmpegRunner(),
